+++ apps/io-func-sign-issuer/src/infra/http/handlers/info.ts
@@ -55,8 +55,7 @@ type InfoDependencies = {
   pdvTokenizerClient: PdvTokenizerClientWithApiKey;
   ioApiClient: IOApiClient;
   db: Database;
-  eventHubBillingClient: EventHubProducerClient;
-  eventHubAnalyticsClient: EventHubProducerClient;
+  signEventsClient: EventHubProducerClient;
   uploadedContainerClient: ContainerClient;
   validatedContainerClient: ContainerClient;
   onSignatureRequestReadyQueueClient: QueueClient;
@@ -70,8 +69,7 @@ export const InfoHandler = H.of((_req: H.HttpRequest) =>
         pdvTokenizerClient,
         ioApiClient,
         db,
-        eventHubBillingClient,
-        eventHubAnalyticsClient,
+        signEventsClient,
         uploadedContainerClient,
         validatedContainerClient,
         onSignatureRequestReadyQueueClient
@@ -81,8 +79,7 @@ export const InfoHandler = H.of((_req: H.HttpRequest) =>
             makePdvTokenizerHealthCheck(pdvTokenizerClient)(),
             makeIOServicesHealthCheck(ioApiClient)(),
             makeAzureCosmosDbHealthCheck(db),
-            makeAzureEventHubHealthCheck(eventHubBillingClient),
-            makeAzureEventHubHealthCheck(eventHubAnalyticsClient),
+            makeAzureEventHubHealthCheck(signEventsClient),
             makeAzureStorageContainerHealthCheck(uploadedContainerClient),
             makeAzureStorageContainerHealthCheck(validatedContainerClient),
             makeAzureStorageQueueHealthCheck(onSignatureRequestReadyQueueClient)