+++ infra/italynorth/_modules/events_function_app/main.tf
@@ -28,7 +28,18 @@ module "function_sign_events" {
 
   app_settings = local.io_sign_events_func.app_settings
 
-  slot_app_settings = local.io_sign_events_func.app_settings
+  slot_app_settings = merge(
+    local.io_sign_events_func.app_settings,
+    {
+      "AzureWebJobs.signEventTrigger_pdnd.Disabled"    = "1"
+      "AzureWebJobs.signEventTrigger_webhook.Disabled" = "1"
+    }
+  )
+
+  sticky_app_setting_names = [
+    "AzureWebJobs.signEventTrigger_pdnd.Disabled",
+    "AzureWebJobs.signEventTrigger_webhook.Disabled",
+  ]
 
   action_group_ids = [data.azurerm_monitor_action_group.common_error_action_group.id, data.azurerm_monitor_action_group.sign_error_action_group.id]