+++ apps/io-func-sign-issuer/CHANGELOG.md
@@ -1,5 +1,22 @@
 # io-func-sign-issuer
 
+## 1.10.0
+
+### Minor Changes
+
+- 1bf4b00: Route all issuer events through the sign-events pipeline
+
+## 1.9.0
+
+### Minor Changes
+
+- ca6a9af: Switch createSignatureRequest ingestion from analytics to sign-events
+
+### Patch Changes
+
+- Updated dependencies [ca6a9af]
+  - @io-sign/io-sign@1.12.0
+
 ## 1.8.6
 
 ### Patch Changes