+++ apps/io-sign-backoffice-app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # io-sign-backoffice-app
 
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies [ca6a9af]
+  - @io-sign/io-sign@1.12.0
+
 ## 2.0.0
 
 ### Major Changes