+++ apps/io-sign-backoffice-func/CHANGELOG.md
@@ -1,5 +1,12 @@
 # io-sign-backoffice-func
 
+## 1.8.5
+
+### Patch Changes
+
+- Updated dependencies [ca6a9af]
+  - @io-sign/io-sign@1.12.0
+
 ## 1.8.4
 
 ### Patch Changes